Rename myController.js to index.js
